requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==7.1.3
python-dotenv==0.19.0
python-telegram-bot==13.7
requests==2.26.0
telegram~=0.0.1